Cuda GPU integration via loading compiled PTX files and baseband beamformer (#1068)
This adds the ability to load compiled PTX kernel for NVIDIA/CUDA GPUs.

Adds the baseband beamforming PTX kernel for the CHORD project.

As a major side-effect, we also changed the buffer size type from int to size_t, so this produced a cascade of type changes throughout!

+++ lib/core/buffer.c
@@ -60,7 +60,7 @@ void private_reset_consumers(struct Buffer* buf, const int ID);
  */
 int private_mark_frame_empty(struct Buffer* buf, const int id);
 
-struct Buffer* create_buffer(int num_frames, int len, struct metadataPool* pool,
+struct Buffer* create_buffer(int num_frames, size_t len, struct metadataPool* pool,
                              const char* buffer_name, const char* buffer_type, int numa_node) {
 
     assert(num_frames > 0);
